% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IV_4K.R
\docType{data}
\name{IV_4K}
\alias{IV_4K}
\title{Damp Heat Plus Dynamic Mechanical Load Indoor Accelerated Test I-V Curve.}
\format{
A data frame with 3637 rows and 2 variables:
\describe{
  \item{V}{increasing vector}
  \item{I}{decreasing vector}
}
}
\source{
Solar Durability and Lifetime Extension (SDLE) Research Center, Case Western
Reserve University
}
\usage{
data(IV_4K)
}
\description{
The I-V curve data is provided by UL for MLEET (Module Level Exposure and Evaluation Test) project funded by DOE (Department of Energy). It was taken from some commercial brand full size PV (photovoltaic) module under dampheat + DML (Dynamic Mechanical Load) indoor accelerated test. Each I-V curve has about 3800 points.
This data set is a study of IV curve with 1 step. The first column is voltage(V) and the second column is current(I).
}
\keyword{datasets}
