% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IV_5M_1.R
\docType{data}
\name{IV_5M_1}
\alias{IV_5M_1}
\title{I-V Curves from External Solar Testing Laboratory.}
\format{
A data frame with 478 rows and 2 variables:
\describe{
  \item{V}{increasing vector}
  \item{I}{decreasing vector}
}
}
\source{
Solar Durability and Lifetime Extension (SDLE) Research Center, Case Western
Reserve University
}
\usage{
data(IV_5M_1)
}
\description{
These 2 IV curves are from an external solar module testing facility. 2 full sized modules are featured, a standard poly-Si AL-BSF module and a similar PERC mono-Si module. The PERC module has a clearly higher power output compared to the Al-BSF module given the higher voltage and current output with a similar fill factor.
This data set is a study of IV curve with 1 step. The first column is voltage(V) and the second column is current(I).
}
\keyword{datasets}
