% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IV_5M_2.R
\docType{data}
\name{IV_5M_2}
\alias{IV_5M_2}
\title{I-V Curves from External Solar Testing Laboratory.}
\format{
A data frame with 476 rows and 2 variables:
\describe{
  \item{V}{increasing vector}
  \item{I}{decreasing vector}
}
}
\source{
Solar Durability and Lifetime Extension (SDLE) Research Center, Case Western
Reserve University
}
\usage{
data(IV_5M_2)
}
\description{
This data set is a study of IV curve with 1 step. The first column is voltage(V) and the second column is current(I).
}
\keyword{datasets}
