% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IV_step1.r
\docType{data}
\name{IV_step1}
\alias{IV_step1}
\title{A data frame of IV curve with 1 step.}
\format{
A data frame with 41 rows and 2 variables:
\describe{
  \item{V}{increasing vector}
  \item{I}{decreasing vector}
}
}
\source{
Solar Durability and Lifetime Extension (SDLE) Research Center, Case Western
Reserve University
}
\usage{
data(IV_step1)
}
\description{
This data set is a study of IV curve with 1 step. The first column is voltage(V) and the second column is current(I).
}
\keyword{datasets}
